feat: Add view pinned messages in room feature
- Added the `getPinnedMessagesByRoomId` function in the backend to fetch
pinned messages. - Implemented the corresponding controller function in
the `roomController`. - Added a new route in `roomRoutes.js` to get
pinned messages. - Integrated the new API in the frontend to display
pinned messages in `PinnedMessages` component. - Added a button in
`ChatArea` to toggle the visibility of pinned messages. - Enhanced
message display with pinned status in `MessageList`. - Added
`getPinnedMessages` API call to `ApiService`.